+++ qiskit/ignis/logging/__init__.py
@@ -1,0 +1,5 @@
+"""
+Ignis Logging package initialization
+"""
+
+from .ignis_logging import IgnisLogger, IgnisLogging, IgnisLogReader