+++ qiskit/ignis/verification/tomography/basis/tomographybasis.py
@@ -1,17 +1,23 @@
 # -*- coding: utf-8 -*-
 
-# Copyright 2018, IBM.
+# This code is part of Qiskit.
 #
-# This source code is licensed under the Apache License, Version 2.0 found in
-# the LICENSE.txt file in the root directory of this source tree.
+# (C) Copyright IBM 2019.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
 
 """
 TomographyBasis class
 """
 
-from qiskit import QuantumRegister
-from qiskit import ClassicalRegister
 from qiskit import QiskitError
+from qiskit.circuit import Qubit, Clbit
 
 
 class TomographyBasis:
@@ -78,12 +84,10 @@ class TomographyBasis:
             raise QiskitError(
                 "{} is not a measurement basis".format(self._name))
 
-        if not (isinstance(qubit, tuple) and isinstance(qubit[0],
-                                                        QuantumRegister)):
+        if not isinstance(qubit, Qubit):
             raise QiskitError('Input must be a qubit in a QuantumRegister')
 
-        if not (isinstance(clbit, tuple) and isinstance(clbit[0],
-                                                        ClassicalRegister)):
+        if not isinstance(clbit, Clbit):
             raise QiskitError('Input must be a bit in a ClassicalRegister')
 
         if op not in self._measurement_labels:
@@ -101,8 +105,7 @@ class TomographyBasis:
             raise QiskitError("{} is not a preparation basis".format(
                 self._name))
 
-        if not (isinstance(qubit, tuple) and isinstance(qubit[0],
-                                                        QuantumRegister)):
+        if not isinstance(qubit, Qubit):
             raise QiskitError('Input must be a qubit in a QuantumRegister')
 
         if op not in self._preparation_labels: